[RESTful 게시판] 서비스, 게시글 서비스 로직 분리
