+++ RESTFULBOARD/src/main/java/board/rest/api/controller/BoardController.java
@@ -1,6 +1,5 @@
 package board.rest.api.controller;
 
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
@@ -15,28 +14,26 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.util.UriUtils;
 
 import board.rest.api.dto.BoardDto;
 import board.rest.api.dto.BoardFileDto;
+import board.rest.api.service.BoardFileService;
 import board.rest.api.service.BoardService;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
 
 @Controller
 @RequiredArgsConstructor
-@Slf4j
-@RequestMapping("/board")
 public class BoardController {
 	
 	private final BoardService boardService;
+	private final BoardFileService boardFileService;
 	
-	// 게시판 목록 조회 화면
-	@GetMapping("")
+	/** 게시판 목록 조회 화면 */
+	@GetMapping("/board")
 	public String openBoardList(Model model) throws Exception{
 		List<BoardDto> list = boardService.selectBoardList();
 		
@@ -45,36 +42,61 @@ public class BoardController {
 		return "/board/restBoardList";		
 	}
 	
-	// 게시글 등록 화면
-	@GetMapping("/write")
+	/** 게시글 등록 화면 */
+	@GetMapping("/board/view")
 	public String openBoardWrite(Model model) throws Exception{
 		return "/board/restBoardWrite";
 	}
 	
-	// 게시글 등록 처리(+ 첨부파일 업로드)
-	@PostMapping("/write")
+	/** 게시글 등록 처리(+ 첨부파일 업로드) */
+	@PostMapping("/board")
 	public String insertBoard(BoardDto board, @RequestParam("files") List<MultipartFile> files) throws Exception{
 	    boardService.insertBoard(board, files);
 	    
 	    return "redirect:/board";
 	}
 	
-	// 게시글 상세 조회 화면(+ 첨부파일 조회)
-	@GetMapping("/{boardIdx}")
+	/** 게시글 상세 조회 화면(+ 첨부파일 조회) */
+	@GetMapping("/board/{boardIdx}")
 	public String openBoardDetail(@PathVariable("boardIdx") int boardIdx, Model model) throws Exception{
 		// @PathVariable : 요청(Request)을 받아와서 리소스에 사용 가능
 		// ("매핑에 사용할 변수 이름") 변수와 묶어 줄 매개변수
-		BoardDto board = boardService.selectBoardDetail(boardIdx);
+		BoardDto board = boardService.selectBoardDetail(boardIdx);		
 		
 		model.addAttribute("board", board);
 		
 		return "/board/restBoardDetail";
+	}	
+	
+	/** 게시글 수정 화면 */
+	@GetMapping("/board/view/{boardIdx}")
+	public String openBoardModify(@PathVariable("boardIdx") int boardIdx, Model model) throws Exception{
+		BoardDto board = boardService.selectBoardDetail(boardIdx);
+		model.addAttribute("board", board);
+		
+		return "/board/restBoardModify";
 	}
 	
-	// 첨부파일 다운로드
-	@GetMapping("/file")
+	/** 게시글 수정 처리 */
+	@PutMapping("/board/{boardIdx}")
+	public String modifyBoard(@PathVariable("boardIdx") int boardIdx, BoardDto board) throws Exception{
+		boardService.modifyBoard(board);
+		
+		return "redirect:/board/" + boardIdx;
+	}
+	
+	/** 게시글 삭제 처리 */
+	@DeleteMapping("/board/{boardIdx}")
+	public String deleteBoard(@PathVariable("boardIdx") int boardIdx) throws Exception{
+		boardService.deleteBoard(boardIdx);
+		
+		return "redirect:/board";
+	}
+	
+	/** 첨부파일 다운로드 */
+	@GetMapping("/board/file")
 	public void downloadBoardFile(BoardFileDto file, HttpServletResponse response) throws Exception{
-		BoardFileDto boardFile = boardService.selectBoardFileInfo(file);
+		BoardFileDto boardFile = boardFileService.selectBoardFileInfo(file);
 		
 		//boardFile은 따로 만든 dto 클래스이기 때문에 MultipartFile처럼 isEmpty 메서드가 없으니까 ObjectUitls.isEmpty메서드 사용
 		if(!ObjectUtils.isEmpty(boardFile)) {
@@ -121,31 +143,4 @@ public class BoardController {
 			outStream.close();  
 		}
 	}
-	
-	
-	// 게시글 수정 화면
-	@GetMapping("/modify/{boardIdx}")
-	public String openBoardModify(@PathVariable("boardIdx") int boardIdx, Model model) throws Exception{
-		BoardDto board = boardService.selectBoardDetail(boardIdx);
-		model.addAttribute("board", board);
-		
-		return "/board/restBoardModify";
-	}
-	
-	// 게시글 수정 처리
-	@PutMapping("/modify/{boardIdx}")
-	public String modifyBoard(@PathVariable("boardIdx") int boardIdx, BoardDto board) throws Exception{
-		boardService.modifyBoard(board);
-		
-		return "redirect:/board/" + boardIdx;
-	}
-	
-	// 게시글 삭제 처리
-	@DeleteMapping("/delete/{boardIdx}")
-	public String deleteBoard(@PathVariable("boardIdx") int boardIdx) throws Exception{
-		boardService.deleteBoard(boardIdx);
-		
-		return "redirect:/board";
-	}
-	
 }
