+++ RESTFULBOARD/src/main/resources/templates/board/restBoardDetail.html
@@ -8,7 +8,7 @@
 <body>
 	<div class="container">
 		<h2>게시글 상세 화면</h2>
-	    <form id="frm" th:method="delete" th:action="'/board/delete/' + ${boardIdx}">
+	    <form id="frm" th:method="delete" th:action="'/board/' + ${boardIdx}">
 	    	<table class="board_detail">
 				<colgroup>
 					<col width="15%"/>
@@ -43,11 +43,12 @@
 			</table>
 			<input type="hidden" id="boardIdx" name="boardIdx" th:value="${board.boardIdx}">
 			<input type="hidden" id="method" name="_method"/>
-			<div class="file_list" th:if="${board.fileList != null  and board.fileList.size() > 0}">[ 첨부파일 ]
+			<div class="file_list" th:if="${board.fileList != null  and board.fileList.size() > 0}">
+				[ 첨부파일 ]
 				<a th:each="list : ${board.fileList}" th:href="@{/board/file(idx=${list.idx}, boardIdx=${list.boardIdx})}" th:text="|${list.originalFileName} (${list.fileSize} kb)|"></a>
 			</div>	
-			<a href="/board" id="list" class="btn">목록으로</a>
-			<a th:href="'/board/modify/' + ${boardIdx}" id="edit" class="btn">수정하기</a>
+			<a href="/board" id="list" class="btn btn_a">목록으로</a>
+			<a th:href="'/board/view/' + ${boardIdx}" id="edit" class="btn btn_a">수정하기</a>
 			<button type="submit" id="delete" class="btn">삭제하기</button>
 		</form>		
 	</div>
