+++ RESTFULBOARD/src/main/resources/templates/board/restBoardList.html
@@ -41,7 +41,7 @@
 	        	</tr>
 	        </tbody>
 	    </table>
-	    <a href="/board/write" class="btn">글 쓰기</a>
+	    <a href="/board/view" class="btn btn_a">글 쓰기</a>
     </div>
 </body>
 </html>
