+++ RESTFULBOARD/src/main/resources/templates/board/restBoardWrite.html
@@ -9,7 +9,7 @@
 	<div class="container">
 	    <h2>게시글 등록</h2>
 	   <!-- enctype="multipart/form-data"를 추가해주는 것이 포인트 -->
-		<form id="frm" name="frm" method="post" action="/board/write" enctype="multipart/form-data">
+		<form id="frm" name="frm" method="post" action="/board" enctype="multipart/form-data">
 			<table class="board_detail">
 	    		<tr>
 	    			<td>제목</td>
